Refactor: Replace GlobalScope pending-image cleanup with ApplicationScope
- Created an `ApplicationScope` object in `util/` providing a `SupervisorJob + Dispatchers.IO` coroutine scope for safe, application-level background tasks.
- Extracted the file deletion logic into a generic suspend `deleteFiles` helper within `FileUtils.kt` in `util/`.
- Updated `CreateVoiceActivity.onDestroy` to use `ApplicationScope.io.launch` and the new helper instead of `GlobalScope`.
- Dropped the `@OptIn(DelicateCoroutinesApi::class)` annotation from `CreateVoiceActivity`.
- Relocated and renamed the unit test to `FileUtilsTest.kt` to ensure the isolated deletion logic behaves correctly.
