+++ app/src/main/java/org/ole/planet/myplanet/lite/util/ApplicationScope.kt
@@ -1,0 +1,9 @@
+package org.ole.planet.myplanet.lite.util
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+
+object ApplicationScope {
+    val io = CoroutineScope(SupervisorJob() + Dispatchers.IO)
+}
