+++ app/src/main/java/org/ole/planet/myplanet/lite/util/FileUtils.kt
@@ -1,0 +1,18 @@
+package org.ole.planet.myplanet.lite.util
+
+import java.io.File
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+object FileUtils {
+    suspend fun deleteFiles(files: List<File>, dispatcher: CoroutineDispatcher = Dispatchers.IO) {
+        withContext(dispatcher) {
+            files.forEach { file ->
+                if (file.exists()) {
+                    file.delete()
+                }
+            }
+        }
+    }
+}
